Adapted and applied jack_lsp filter patch by Ken Restivo.
Original message:

jack_lsp -c  is great for showing all connections, but I needed a way to to filter out just those that were connected to a particular 
device.

"grep -v" isn't sufficient when using "jack_lsp -c", because the connected devices aren't on the same line as what tey are connected 
to. So, rather than make the shell script really complicated, or the output of jack_lsp really ugly, I patched jack_lsp to add a 
"filter" feature. 

I needed this so that I could write a shell script that does "jack_lsp -c alsa_pcm:playback_2" to get a list of everything that is 
connected to that port, and then it can jack_disconnect those (or some of those) and then jack_connect them to another device.

I doubt that anyone else would need this feature, but, just in case, here it is.

-ken



git-svn-id: svn+ssh://jackaudio.org/trunk/jack@1074 0c269be4-1314-0410-8aa9-9f06e86f4224