fix arguably incorrect marking of client with errors after process cycle fails to return to server
When reading the byte from the wait file descriptor failed, jack would mark the client at the head of the external client chain
with an error. But the fault may not have been with that client. So now if the read error is EAGAIN ('no data available') do
not mark the client, but rely on other mechanisms to detect too-slow or dead clients.

Also tweak VERBOSE output a little in some related spots.