Patch from Fons Adriaensen to improve validity/usability of jack_frame_time()
 ensures that frame time will reflect any
discontinuity in processing (xrun or skipped cycles). It also cleans up
the code related to maintaining the DLL and frame time a bit, all of it
is now together in jack_run_cycle() and the logic behind it is easy to
follow.

The 'delayed_usecs' argument to jack_run_cycle() is now probably
redundant, the value computed locally (the 'dus' variable) is in
all cases more accurate than what the backend can provide.