Revert "Set driver port order metadata".
The "client" argument being passed to jack_set_property() is
jack_client_internal_t*, not jack_client_t* as expected, which leads
to a crash. Using client->private_client (which is a jack_client_t*)
causes deadlock because do_request() is called recursively.

The goal here is good, but it can't be accomplished in this way.

This reverts commit 439e47da1a179f62ce9e6601851b4f90381da0de.