configure: update -mmacosx-version-min= to 10.6
This could be done other ways if it turns out there's a flood of
people out there desperate for 10.4 or 10.5 compatibility, but I
expect this is safe enough to solve in the simplest possible way,
which is by just bumping the number here to 10.6.

Fixes https://github.com/jackaudio/jack1/issues/81.